User generation (#8)
We laid in the DB schema for users previously. Now we generate them. Only generation and testing the generation, no doing things. Note that the actual authn will be with respect to a fkey relation to the user - hell, we might even have a join table eventually, that'll be fun.