Lot of notes as to login impl, ci (#17)
We need CI but nowhere to actually deploy yet, but we should use a CI/CD thing. Github actions is kind of remarkably janky and goes down every month like clockwork but it's free for us and straightforward.

There are also a lot of notes as to things. None of the putative revisions to usage.md yet