Authn and trivial authz behind feature flag (#22)
It auths. If you don't auth things go boom. The sessions are in DB which is annoying but I'm not going to get peeps to install redis or whatever for a while yet, and JWT tokens are misery.

Still no FE work. Killing the feature flag after FE work done